+++ src/systems/mod.rs
@@ -1,8 +1,0 @@
-pub mod camera;
-pub mod input;
-pub mod loading;
-pub mod physics;
-pub mod simulation_actions;
-#[allow(clippy::type_complexity)]
-pub mod ui;
-pub mod visualization;